refactor: extract create task modal
